Sth improved in scrolling and cursor positioning.

+++ src/fiflo.h
@@ -8,7 +8,7 @@
 #include <stdint.h>
 
 typedef uint16_t term_t;
-typedef int8_t buff_t; // TEMPONARY.
+typedef int16_t buff_t; // TEMPONARY.
 
 #define MAX_LNS (buff_t) powf(2, (sizeof(buff_t) * 8))
 #define MAX_CHRS MAX_LNS - 1 // 1 for NULL. TODO
